Refactor PMA managed thread route helpers
